refactor(ui): extract CourseTableView and ScoreListTab/ScoreAnalysisTab as public widgets
Extract course table grid rendering into CourseTableView and score
tab widgets into ScoreListTab/ScoreAnalysisTab for reuse by the
Liquid Glass package. Original scaffolds delegate to the new widgets.
