+++ packages/ap_common_flutter_ui/lib/ap_common_flutter_ui.dart
@@ -26,6 +26,7 @@ export 'src/widgets/about_info_card.dart';
 export 'src/widgets/ap_drawer.dart';
 export 'src/widgets/ap_network_image.dart';
 export 'src/widgets/color_picker_widgets.dart';
+export 'src/widgets/course_table_view.dart';
 export 'src/widgets/default_dialog.dart';
 export 'src/widgets/dialog_option.dart';
 export 'src/widgets/hint_banner.dart';
@@ -37,6 +38,7 @@ export 'src/widgets/option_dialog.dart';
 export 'src/widgets/phone_list_item.dart';
 export 'src/widgets/progress_dialog.dart';
 export 'src/widgets/score_analysis_widgets.dart';
+export 'src/widgets/score_tab_widgets.dart';
 export 'src/widgets/semester_picker.dart';
 export 'src/widgets/setting_page_widgets.dart';
 export 'src/widgets/yes_no_dialog.dart';
